+++ src/BundlerMinifierVsix/Constants.cs
@@ -2,7 +2,7 @@
 {
     class Constants
     {
-        public const string FILENAME = "bundleconfig.json";
+        public const string CONFIG_FILENAME = "bundleconfig.json";
         public const string VSIX_NAME = "Bundler & Minifier";
         public const string NUGET_ID = "BuildBundlerMinifier";
     }