+++ src/BundlerMinifierVsix/FileListeners/JsonCreationListener.cs
@@ -30,7 +30,7 @@ namespace BundlerMinifierVsix.Listeners
             {
                 string fileName = Path.GetFileName(_document.FilePath);
 
-                if (fileName.Equals(Constants.FILENAME, StringComparison.OrdinalIgnoreCase))
+                if (fileName.Equals(Constants.CONFIG_FILENAME, StringComparison.OrdinalIgnoreCase))
                 {
                     _document.FileActionOccurred += DocumentSaved;
                 }