+++ src/BundlerMinifierVsix/JSON/RelativeFilePathFormatProvider.cs
@@ -27,7 +27,7 @@ namespace BundlerMinifierVsix.JSON
 
             string fileName = Path.GetFileName(doc.DocumentLocation);
 
-            if (!fileName.Equals(Constants.FILENAME, StringComparison.OrdinalIgnoreCase))
+            if (!fileName.Equals(Constants.CONFIG_FILENAME, StringComparison.OrdinalIgnoreCase))
                 yield break;
 
             string folder = Path.GetDirectoryName(doc.DocumentLocation);