+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Roadmap
 
-- [ ] Task Runner Explorer integration
+- [x] Task Runner Explorer integration
+- [x] Command line support
 - [ ] Use FileSystemWatcher (#28)
 
 Features that have a checkmark are complete and available for